transferred to api folder for specific test case per endpoint
